Review vuex module refactored
