+++ core/modules/review/store/actions.ts
@@ -2,10 +2,10 @@ import Vue from 'vue'
 import { ActionTree } from "vuex"
 import { quickSearchByQuery } from "core/store/lib/search"
 import SearchQuery from 'core/store/lib/search/searchQuery'
-import { adjustMultistoreApiUrl } from '../../lib/multistore'
-import RootState from "../../types/RootState"
-import ReviewState from "./types/ReviewState"
-import * as types from '../../mutation-types'
+import { adjustMultistoreApiUrl } from '@vue-storefront/store/lib/multistore'
+import RootState from "@vue-storefront/store/types/RootState"
+import ReviewState from "../types/ReviewState"
+import * as types from './mutation-types'
 import i18n from '@vue-storefront/i18n'
 import rootStore from "core/store"
 import { ValidationError } from "core/store/lib/exceptions";
