+++ core/modules/review/store/index.ts
@@ -1,8 +1,8 @@
 import { Module } from 'vuex'
 import actions from './actions'
 import mutations from "./mutations";
-import RootState from "../../types/RootState";
-import ReviewState from "./types/ReviewState";
+import RootState from "@vue-storefront/store/types/RootState";
+import ReviewState from "../types/ReviewState";
 
 const review: Module<ReviewState, RootState> = {
   namespaced: true,
