+++ core/modules/review/store/mutation-types.ts
@@ -1,0 +1,2 @@
+export const SN_REVIEW = 'review'
+export const REVIEW_UPD_REVIEWS = SN_REVIEW + 'UPD_REVIEWS'
