+++ core/modules/review/store/mutations.ts
@@ -1,6 +1,6 @@
 import { MutationTree } from 'vuex'
-import * as types from '../../mutation-types'
-import ReviewState from './types/ReviewState'
+import * as types from './mutation-types'
+import ReviewState from '../types/ReviewState'
 
 const mutations: MutationTree<ReviewState> = {
   [types.REVIEW_UPD_REVIEWS] (state, items) {
