+++ core/modules/review/store/review.schema.json
